Modify The Core API To Run Arbitrary Effects
Some awkwardness was discovered in having the `primitive` API method encapsulate a state value directly. This commit changes the core API to merely call two effects at the start and end of the http transaction.

This commit also adds an `onMax` effect to the higher level APIs which is invoked when the number of concurrent requests exceeds the specified limit. While the polling effect is still useful for use cases such as metrics, it can not be used to know the _exact_ moment we go over the limit due to the concurrent nature of the value.